tests: plots: reorginize the structure
