--- tests/func/metrics/conftest.py
@@ -1,1 +1,0 @@
-from tests.func.plots import run_copy_metrics  # noqa: F401
