+++ tests/func/plots/conftest.py
@@ -1,1 +1,12 @@
-from . import run_copy_metrics  # noqa: F401
+import shutil
+
+import pytest
+
+
+@pytest.fixture()
+def custom_template(tmp_dir, dvc):
+    custom_template = tmp_dir / "custom_template.json"
+    shutil.copy(
+        tmp_dir / ".dvc" / "plots" / "default.json", custom_template,
+    )
+    return custom_template
