+++ tests/func/plots/test_plots.py
@@ -9,7 +9,6 @@ import pytest
 import yaml
 from funcy import first
 
-from dvc.dvcfile import PIPELINE_LOCK
 from dvc.repo.plots.data import (
     JSONPlotData,
     NoMetricInHistoryError,
@@ -18,7 +17,6 @@ from dvc.repo.plots.data import (
     YAMLPlotData,
 )
 from dvc.repo.plots.template import NoFieldInDataError, TemplateNotFoundError
-from dvc.utils import relpath
 
 
 def _write_csv(metric, filename, header=True):
@@ -358,15 +356,6 @@ def test_throw_on_no_metric_at_all(tmp_dir, scm, dvc, caplog):
     assert str(error.value) == "Could not find 'metric.json'."
 
 
-@pytest.fixture()
-def custom_template(tmp_dir, dvc):
-    custom_template = tmp_dir / "custom_template.json"
-    shutil.copy(
-        tmp_dir / ".dvc" / "plots" / "default.json", custom_template,
-    )
-    return custom_template
-
-
 def test_custom_template(tmp_dir, scm, dvc, custom_template, run_copy_metrics):
     metric = [{"a": 1, "b": 2}, {"a": 2, "b": 3}]
     _write_json(tmp_dir, metric, "metric_t.json")
@@ -540,51 +529,6 @@ def test_load_metric_from_dict_yaml(tmp_dir):
         d["rev"] = "revision"
 
     assert list(map(dict, plot_data.to_datapoints())) == expected
-
-
-def test_plots_modify_existing_template(
-    tmp_dir, dvc, run_copy_metrics, custom_template
-):
-    metric = [{"a": 1, "b": 2}, {"a": 2, "b": 3}]
-    _write_json(tmp_dir, metric, "metric_t.json")
-    stage = run_copy_metrics(
-        "metric_t.json",
-        "metric.json",
-        plots_no_cache=["metric.json"],
-        name="copy-metrics",
-        single_stage=False,
-    )
-    dvc.plots.modify(
-        "metric.json", props={"template": relpath(custom_template)}
-    )
-    stage = stage.reload()
-    assert stage.outs[0].plot == {"template": relpath(custom_template)}
-
-
-def test_plots_modify_should_not_change_lockfile(
-    tmp_dir, dvc, run_copy_metrics, custom_template
-):
-    _write_json(tmp_dir, [{"a": 1, "b": 2}], "metric_t.json")
-    run_copy_metrics(
-        "metric_t.json",
-        "metric.json",
-        plots_no_cache=["metric.json"],
-        name="copy-metrics",
-        single_stage=False,
-    )
-
-    (tmp_dir / PIPELINE_LOCK).unlink()
-    dvc.plots.modify(
-        "metric.json", props={"template": relpath(custom_template)}
-    )
-    assert not (tmp_dir / PIPELINE_LOCK).exists()
-
-
-def test_plots_modify_not_existing_template(dvc):
-    with pytest.raises(TemplateNotFoundError):
-        dvc.plots.modify(
-            "metric.json", props={"template": "not-existing-template.json"}
-        )
 
 
 def test_multiple_plots(tmp_dir, scm, dvc, run_copy_metrics):
